added toast and removed alert
